Minor fix to ensure logs are kept for test failure.

+++ tests/ship_test.py
@@ -208,7 +208,7 @@ finally:
                 # output file should have lots of output, but if user passes in a huge number of requests, these could go on forever
                 printTruncatedFile("%s%d.out" % (shipClientFilePrefix, i), maxLines=20000)
 
-        if not keepLogs:
+        if testSuccessful and not keepLogs:
             shutil.rmtree(shipTempDir, ignore_errors=True)
     if not testSuccessful and dumpErrorDetails:
         Print(Utils.FileDivider)
